+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to the PDF Watermark Remover will be documented in this file.
 
+## [1.0.1] - 2025-04-03
+
+### Security Improvements
+- Use constant-time comparison.
+- Strictly verify SSL certificates.
+- Ensured environment variables 
+
 ## [1.0.0] - 2025-03-25
 
 ### Added